requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
Django==4.1.2
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
drf-extensions==0.7.1
drf-nested-routers==0.93.4
execnet==1.9.0
freezegun==1.2.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.4
routers==0.10.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
